fix(ml): fix pickle IsotonicCalibratedXGB y label encoder BERT
- _calibrated_model.py: mueve IsotonicCalibratedXGB a módulo propio para
  que joblib pueda deserializarla correctamente (evita __main__ pickle error)
- calibrate.py: importa IsotonicCalibratedXGB desde _calibrated_model
- bert_pipeline/models/label_encoder.joblib: copiado desde modelo fusionado
  (mismas 4 clases en mismo orden, verificado con config.json de BERT)

Ensemble XGBoost+BERT completamente operativo: xgboost_bert activo.
